requirements.txt
fastapi
uvicorn
jinja2
python-multipart
aiofiles
itsdangerous
openpyxl
python-dotenv
pandas
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
psycopg2-binary
